requirements.txt
py-clob-client>=0.29.0
anthropic>=0.40.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
apscheduler>=3.10.0
structlog>=24.0.0
web3==6.14.0
